+++ NEWS.md
@@ -2,6 +2,10 @@
 
 # What's new in SITS version 1.1.0
 
+### Hotfix version 1.1.0-2
+* Improve sits_values() function (issue #810)
+* Fix sits_reduce_imbalance() function (issue #809)
+
 ### Hotfix version 1.1.0-1
 * Fix sits_accuracy() function (issue #807)